refactor AutoRefillModule and AutoToolModule to improve inventory handling and add utility functions
