+++ src/main/kotlin/org/xodium/vanillaplus/modules/InvUnloadModule.kt
@@ -10,12 +10,15 @@ import io.papermc.paper.command.brigadier.CommandSourceStack
 import io.papermc.paper.command.brigadier.Commands
 import net.kyori.adventure.sound.Sound
 import org.bukkit.NamespacedKey
+import org.bukkit.Tag
 import org.bukkit.block.Block
 import org.bukkit.block.Container
+import org.bukkit.block.ShulkerBox
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
 import org.bukkit.event.player.PlayerQuitEvent
+import org.bukkit.inventory.Inventory
 import org.xodium.vanillaplus.Config
 import org.xodium.vanillaplus.Perms
 import org.xodium.vanillaplus.VanillaPlus.Companion.instance
@@ -75,7 +78,7 @@ class InvUnloadModule : ModuleInterface {
         val affectedChests = mutableListOf<Block>()
         for (block in useableChests) {
             val inv = (block.state as Container).inventory
-            if (Utils.stuffInventoryIntoAnother(player, inv, true, startSlot, endSlot)) {
+            if (stuffInventoryIntoAnother(player, inv, true, startSlot, endSlot)) {
                 affectedChests.add(block)
             }
         }
@@ -96,4 +99,49 @@ class InvUnloadModule : ModuleInterface {
 
         player.playSound(Config.InvUnloadModule.SOUND_ON_UNLOAD, Sound.Emitter.self())
     }
+
+    /**
+     * Moves items from the player's inventory to another inventory.
+     * @param player The player whose inventory is being moved.
+     * @param destination The destination inventory to move items into.
+     * @param onlyMatchingStuff If true, only moves items that match the destination's contents.
+     * @param startSlot The starting slot in the player's inventory to move items from.
+     * @param endSlot The ending slot in the player's inventory to move items from.
+     * @return True if items were moved, false otherwise.
+     */
+    private fun stuffInventoryIntoAnother(
+        player: Player,
+        destination: Inventory,
+        onlyMatchingStuff: Boolean,
+        startSlot: Int,
+        endSlot: Int,
+    ): Boolean {
+        val source = player.inventory
+        val initialCount = countInventoryContents(source)
+        var moved = false
+
+        for (i in startSlot..endSlot) {
+            val item = source.getItem(i) ?: continue
+            if (Tag.SHULKER_BOXES.isTagged(item.type) && destination.holder is ShulkerBox) continue
+            if (onlyMatchingStuff && !Utils.doesChestContain(destination, item)) continue
+
+            val leftovers = destination.addItem(item)
+            val movedAmount = item.amount - leftovers.values.sumOf { it.amount }
+            if (movedAmount > 0) {
+                moved = true
+                source.clear(i)
+                leftovers.values.firstOrNull()?.let { source.setItem(i, it) }
+                destination.location?.let { Utils.protocolUnload(it, item.type, movedAmount) }
+            }
+        }
+        return moved && initialCount != countInventoryContents(source)
+    }
+
+    /**
+     * Counts the total number of items in the given inventory.
+     * @param inv The inventory to count items in.
+     * @return The total number of items in the inventory.
+     */
+    private fun countInventoryContents(inv: Inventory): Int = inv.contents.filterNotNull().sumOf { it.amount }
+
 }
